Tesztesetekhez szukseges dolgok
